feat: reorganize header generation and URL building functions for better structure and maintainability
