+++ src/routes/user.routes.ts
@@ -1,5 +1,5 @@
 import { Elysia } from "elysia";
-import { searchDomain, searchIP } from "@src/controllers/search.controller";
+import { searchDomain, searchIP } from "@src/controllers/searchController/search.controller";
 
 export const appRoutes = new Elysia()
   .get("/search/ip_address/:ip", searchIP, {
