+++ src/services/search/function/generateHeaders/generateHeaders.ts
